Using DBSCAN for the trip clustering as well
We are already using DBSCAN for the start and end location clustering, we can
pretty easily find trips by matching on the start and end locations. Here, we
explore two alternatives for the trip matching - add the distance matrices and
recluster, or group by the (start, end) cluster label pairs and group.

We find that the second method is correct. We also spot check both methods and
find that the location clustering can also have some minor issues sometimes,
but 2/3 spot checks worked well.