Deep dive into the similarity code
Primarily to understand why it is worse than the DBSCAN trip clustering code.

As you may recall, the results for DBSCAN were pretty respectable.
Most trips were in a cluster, and the median cluster : trip-in-cluster ratio
was below 40% for all datasets.
https://github.com/e-mission/e-mission-eval-private-data/pull/28#issuecomment-885958390

Why is similarity so much worse than the DBSCAN-based clustering?
Should we switch to DBSCAN instead (horrifying proposition given the tight deadline)?

See the notebook conclusion to find out!

@corinne-hcr, this is more on the lines of what I expected you to do back when
you were evaluating the first round/common trips, and definitely when we were
getting poor results with the clustering.

+ change the original DBSCAN notebook to have the modified viz code